Add fix.

+++ src/product_settings/chat_widget.rs
@@ -6,7 +6,7 @@ pub struct ChatWidget {
     pub live_chat: Option<ChatSetting>,
     pub zen_desk: Option<ChatSetting>,
     pub hide_widget: Option<bool>,
-    pub chat_type: Option<u8>,
+    pub chat_type: Option<i32>,
 }
 
 #[derive(Serialize, Deserialize, Clone)]
